requirements.txt
telethon
langchain
openai
python-dotenv
